Re-arrange API test helpers
* Moved the ApiHelpers and ApiTestSetup modules into api/testing_support so that applications can use it
* Moved database_cleaner config into spec/support/database_cleaner.rb
