+++ api/spec/support/have_attributes_matcher.rb
@@ -1,0 +1,13 @@
+
+
+
+
+RSpec::Matchers.define :have_attributes do |expected_attributes|
+  match do |actual|
+    # actual is a Hash object representing an object, like this:
+    # { "product" => { "name" => "Product #1" } }
+    actual_attributes = actual.values.first.keys.map(&:to_sym)
+    expected_attributes.map(&:to_sym).all? { |attr| actual_attributes.include?(attr) }
+  end
+end
+
